+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.10.3.0 - 2026-06-27 (NINA 3.3)
 * Fix for new preview report timing issues
+* Added chart to preview report as well as HTML output
 
 ## 5.10.2.0 - 2026-06-26 (NINA 3.3)
 * Added an altitude chart to each selected target in the plan preview.